// frontend/src/openapi/models/ModeCd6Enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `optimize` - Optimize
 * * `model_comparison` - Model Comparison
 * * `hybrid` - Harness + Models
 * @export
 */
export const ModeCd6Enum = {
    optimize: 'optimize',
    model_comparison: 'model_comparison',
    hybrid: 'hybrid'
} as const;
export type ModeCd6Enum = typeof ModeCd6Enum[keyof typeof ModeCd6Enum];


export function instanceOfModeCd6Enum(value: any): boolean {
    for (const key in ModeCd6Enum) {
        if (Object.prototype.hasOwnProperty.call(ModeCd6Enum, key)) {
            if (ModeCd6Enum[key as keyof typeof ModeCd6Enum] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ModeCd6EnumFromJSON(json: any): ModeCd6Enum {
    return ModeCd6EnumFromJSONTyped(json, false);
}

export function ModeCd6EnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModeCd6Enum {
    return json as ModeCd6Enum;
}

export function ModeCd6EnumToJSON(value?: ModeCd6Enum | null): any {
    return value as any;
}

export function ModeCd6EnumToJSONTyped(value: any, ignoreDiscriminator: boolean): ModeCd6Enum {
    return value as ModeCd6Enum;
}