// frontend/src/openapi/models/OptimizerExperiment.ts
export interface OptimizerExperiment {
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly id: string;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly project: string;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly capability: string;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly capabilityName: string;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly datasetName: string;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly evalSetName: string;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly dataset: string | null;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly cell: string | null;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof OptimizerExperiment
     */
    readonly cellInfo: { [key: string]: any; } | null;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly evalSet: string | null;
    /**
     *
     * @type {number}
     * @memberof OptimizerExperiment
     */
    readonly triggeredBy: number | null;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly entrypoint: string;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly codeTrigger: string;
    /**
     *
     * @type {ModeCd6Enum}
     * @memberof OptimizerExperiment
     */
    readonly mode: ModeCd6Enum;
    /**
     *
     * @type {any}
     * @memberof OptimizerExperiment
     */
    readonly modelIds: any | null;
    /**
     *
     * @type {OpenrouterKeySourceEnum}
     * @memberof OptimizerExperiment
     */
    readonly openrouterKeySource: OpenrouterKeySourceEnum;
    /**
     *
     * @type {OptimizerExperimentStatusEnum}
     * @memberof OptimizerExperiment
     */
    readonly status: OptimizerExperimentStatusEnum;
    /**
     *
     * @type {number}
     * @memberof OptimizerExperiment
     */
    readonly currentIteration: number;
    /**
     *
     * @type {number}
     * @memberof OptimizerExperiment
     */
    readonly numIterations: number;
    /**
     *
     * @type {number}
     * @memberof OptimizerExperiment
     */
    readonly numCandidatesPerIteration: number;
    /**
     *
     * @type {number}
     * @memberof OptimizerExperiment
     */
    readonly maxIterationsWithoutImprovement: number;
    /**
     *
     * @type {number}
     * @memberof OptimizerExperiment
     */
    readonly stalledIterations: number;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly commandTemplate: string;
    /**
     *
     * @type {any}
     * @memberof OptimizerExperiment
     */
    readonly scores: any | null;
    /**
     *
     * @type {any}
     * @memberof OptimizerExperiment
     */
    readonly state: any | null;
    /**
     *
     * @type {string}
     * @memberof OptimizerExperiment
     */
    readonly failureReason: string;
    /**
     *
     * @type {Date}
     * @memberof OptimizerExperiment
     */
    readonly createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof OptimizerExperiment
     */
    readonly updatedAt: Date;
}